+++ packages/react-native-sdk/src/hooks/index.ts
@@ -1,4 +1,3 @@
-export * from './useLocalVideoStream';
 export * from './useIncallManager';
 export * from './usePermissionRequest';
 export * from './usePermissionNotification';